requirements.txt
streamlit
moviepy
pydub
openai-whisper
langdetect
